+++ crates/web-client/src/models/word.rs
@@ -1,7 +1,9 @@
 use miden_objects::{Felt as NativeFelt, Word as NativeWord};
 use wasm_bindgen::prelude::*;
+use wasm_bindgen_futures::js_sys::Uint8Array;
 
 use super::felt::Felt;
+use crate::utils::{deserialize_from_uint8array, serialize_to_uint8array};
 
 #[wasm_bindgen]
 #[derive(Clone)]
@@ -43,6 +45,15 @@ impl Word {
     #[wasm_bindgen(js_name = "toHex")]
     pub fn to_hex(&self) -> String {
         self.0.to_hex()
+    }
+
+    pub fn serialize(&self) -> Uint8Array {
+        serialize_to_uint8array(&self.0)
+    }
+
+    pub fn deserialize(bytes: &Uint8Array) -> Result<Word, JsValue> {
+        let native_word = deserialize_from_uint8array::<NativeWord>(bytes)?;
+        Ok(Word(native_word))
     }
 
     #[wasm_bindgen(js_name = "toU64s")]